Add files via upload
Single class module separated in 10 class modules. This makes code easier to maintain. And the usage is more intuitive. Instead of one class with a bunch of properties we now have a hierarchy of classes with less properties each.
It is now possible to specify the source of scripts and css: pivot.js, pivot.css, jquery.js, jquery-ui.js